Rename `traits` to `vector_support`
