+++ libcudacxx/include/cuda/std/detail/libcxx/include/complex
@@ -245,7 +245,7 @@ template<class T> complex<T> tanh (const complex<T>&);
 #endif // no system header
 
 #include <cuda/std/detail/libcxx/include/__assert> // all public C++ headers provide the assertion handler
-#include <cuda/std/detail/libcxx/include/__complex/traits.h>
+#include <cuda/std/detail/libcxx/include/__complex/vector_support.h>
 #include <cuda/std/detail/libcxx/include/__type_traits/enable_if.h>
 #include <cuda/std/detail/libcxx/include/__type_traits/is_constructible.h>
 #include <cuda/std/detail/libcxx/include/__type_traits/is_floating_point.h>
